+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ###### Dev
 
+-   Conditionally use different past artist show fragments for iPad vs iPhone - alloy
+-   Upgrades Relay to ‘modern’ (v1.3.0) - alloy
 -   Upgrades React Native to v0.48.0 - sarah
 -   Extracts connectivity banner into its own component - sarah
 -   Adds stylelint to the dev-experience, not validated on CI yet - orta
@@ -23,6 +25,7 @@
 -   Adds a screen for metadata editing - orta
 -   Looks better on iPads - orta
 -   Adds a screen for setting the location - orta
+-   Adds a spinner at the end of submission - orta
 
 ###### Messaging